refactor(core): move printer source-text helpers into core text domain
Original location: src/format/src/printer/source-text.ts
New location:       src/core/src/text/gml-printer-source-text.ts

Rationale:
The file contained general-purpose GML printer utilities (metadata
resolution, blank-line detection, line-terminator stripping) that did not
depend on Prettier or the formatter's print logic. Moving them into
@gmloop/core's `text` domain aligns with the architectural boundary
between `@gmloop/core` (shared data-model + utilities) and `@gmloop/format`
(layout-focused formatting only).

Specifically:
- stripTrailingLineTerminators: character-level line ending removal
- getOriginalTextFromOptions: option normalization
- resolvePrinterSourceMetadata: type-safe metadata parsing
- resolveNodeIndexRangeWithSource: node range resolution
- sliceOriginalText: safe text slicing
- macroTextHasExplicitTrailingBlankLine: macro text analysis
- hasBlankLineBeforeLeadingComment: blank-line detection
- hasBlankLineBetweenLastCommentAndClosingBrace: block comment analysis

The format workspace's source-text.ts is now a thin shim that imports
from Core and re-exports the helpers, preserving backward compatibility
for relative imports within the printer subsystem.

This keeps layout-focused utilities in the formatter while shared text
utilities live in core, following the target-state.md ownership contract.
