+++ src/core/src/text/index.ts
@@ -1,1 +1,2 @@
+export * from "./gml-printer-source-text.js";
 export * from "./source-text.js";
